+++ CHANGES.rst
@@ -4,6 +4,12 @@ Changelog
 2.5.4 (unreleased)
 ------------------
 
+- Added support for site logos stored in the portal registry via the site
+  control panel for the logo viewlet with a fallback to the ``OFS.Image``
+  based ``logo.png`` file. Removed support of long-gone
+  ``base_properties.props`` defined logo names.
+  [thet]
+
 - Updated markup for dashboard.
   [davisagli]